Got it in a working state
