+++ .gitignore
@@ -1,0 +1,2 @@
+/game-env
+data/bin/__pycache__
