+++ data/bin/inventory.py
