+++ data/sprites/Misc/potion_mana.png
